requirements.txt
fastapi[standard]
requests
scikit-learn
transformers
torch
python-dotenv